Convert demo_qs into test_qs TAP test suite
This converts demo_qs into a TAP test suite and adds the necessary
changes to configure.ac and Makefile.am to run this test suite on
`make check` or `make distcheck`. This adds some initial tests of
the library to ensure that the basic functionality is working.

The `test-driver.sh` script was added even though it is bundled
and automatically installed when running `bootstrap` for newer
version of automake. Unfortunately the Travis CI automake is
too old and this does not work. Also, the Automake 1.11 currently
active on Travis CI does not actually support the log driver at
all so it will happily continue even if TAP tests fail.
